Move i04 beamstop to mx_phase1
